Move version info into certvalidator.version for standalone use
