+++ certvalidator/ocsp_client.py
@@ -8,7 +8,7 @@ from asn1crypto import core, ocsp, x509, algos, util
 
 from . import errors
 from ._types import str_cls, type_name
-from ._version import __version__
+from .version import __version__
 
 if sys.version_info < (3,):
     from urllib2 import Request, urlopen, URLError
